refactor(Explorer): moved etherscan into folder
