+++ test/e2e/verifier/ethereum-main-invalid-merkle-root.test.js
@@ -1,7 +1,7 @@
 import { Certificate, VERIFICATION_STATUSES } from '../../../src';
 import sinon from 'sinon';
 import FIXTURES from '../../fixtures';
-import * as ethereumExplorer from '../../../src/explorers/ethereum';
+import * as ethereumExplorer from '../../../src/explorers/ethereum/etherscan';
 
 describe('given the certificate is an ethereum main with an invalid merkle root', function () {
   let certificate;
